change gcd function to work with int64_t instead of int

+++ Cryptography/gcd.c
@@ -23,7 +23,7 @@ void print_gcd(int64_t a, int64_t b) {
 	printf("%lld = %lld * %lld + %lld * %lld\n", gcd, m, a, n, b);
 }
 
-int gcd(int a, int b) {
+int64_t gcd(int64_t a, int64_t b) {
 	if (a < b)
 		return gcd(b, a);
 	if (a % b == 0)
